Refactor dynload/port.h (#43431)
* Refactor port.h

* Remove some unnecessary code

* Fix CI errors
