+++ paddle/fluid/operators/detection/sigmoid_focal_loss_op.h
@@ -14,6 +14,7 @@ limitations under the License. */
 
 #pragma once
 #include <algorithm>
+#include <cfloat>
 #include <limits>
 
 #include "paddle/fluid/framework/op_registry.h"
